update: 更新文件 bluetooth_ble_advertiser_proxy.cpp

+++ frameworks/inner/ipc/src/bluetooth_ble_advertiser_proxy.cpp
@@ -324,7 +324,7 @@ int BluetoothBleAdvertiserProxy::BleStartRangeAdv(bluetooth::BleAppType appType)
         HILOGE("[BleStartRangeAdv] fail: transact ErrCode=%{public}d", result);
         return INNER_BLE_RANGING_INVALID_ADVPOWER;
     }
-    return reply.ReadInt8();
+    return reply.ReadInt32();
 }
 
 int BluetoothBleAdvertiserProxy::BleStopRangeAdv(bluetooth::BleAppType appType)
